Add navigation bar
Added navigation bar using include-style layouts, meaning that it has
to be inserted into every layout. The benefit of this approach is the
flexibility in developing views, but as the application grows it might
be best to transfer to hierarchical-style, due to the fact that this
solution is somewhat cumbersome.